Added feature to spawn a process and pipe status output to it
When writing status, we don't use stdin but the result from pipe (we
don't dup it) so we can still print loging info normally. Also if there
is no program spawned as a pipe, there the status messages should be
lost.

config.c was modified to show a proof of concept. We intended to
proivide a more feature-full default print method similar to XMonad's at
a later data.

Also we noticed that RUN_OR_RAISE_TITLE was broken so we fixed that

Added spawnPipe to ignore list and refactored said list