Added support for ENV_VARS for Rules (ie when using RUN_OR_RAISE)
Environmental vars can now be used seamlessly with Rules (well almost;
you cannot do anything stupid like combining regex (yet); There may be a
feature later to use any arbitrary bash command)

ENV_VARS are substituted on startup and used as if this string were
literally typed. Unless the string is a REGEX using and the REGEX flag
is pointless.

To accomplish the above, we have to initialize Rules before they are used.
Went ahead an combined this with REGEX; both will be applied when the Rule is first used.

Removed the COMPILE_RULE macro as rule will be compiled on first use.

Updated Custom user config