Fixed double-free bug related to window propties
When loadWindowProperties was called, the previous property was freeded,
it it was set in preperation for the new value. However, if we could not
get a new value, then the property points to freed memory so we
double-freed if loadWindowProperties was called again. This problem only
affects loading the tile as the other properties don't use pointers or
calloc a new one