Hide unmapped windows
Before we checked if a window was activatable to determine if we should
tread it as an active/visible member of the stack (ie for tiling) now we
check if it only if it is mapped. This way when the client unmapps a
window and doesn't add the Hidden property, we don't have blank space in
our tiling nor do we error when cycling focus.