Handle case when client spams requests
Discovered that some clients may (for some reason) bombard us with
requests if we don't grant their requests. This could prevent Idle
events from triggering in a timely mannger. So we added a new event
type that will be triggered after a set number of events or when
Idle is triggered.