Automatic workspace tiling improvement
We now tile workspaces on Periodic events instead of Idle
When tiling workspace we clear the dirty mask so we don't tile/check twice

Add markState hook to be called before keybindings

There was a problem where a keybinding to retile would cause windows to be tiled twice: once for retile, then again for the automatic tiling. The fix for this was to call unMarkState(), but markState was always called after all keybindings so it was pointless. Now markState is called before all keybindings