Added onWindowMove rule
These rules are called anytime a managed window is configured. The filtering out of ignored windows is one of the main differences between this and XCB_CONFIGURE_NOTIFY. THe other being that the WindowInfo object will be passed in when the rule is applied.
This is rule is also triggered on a create after if a window has been
successfully registered