Better window cycling via findAndRaise
Observed problem when cycling through a set of windows where the focused
window would go A B B A A ... instead A B A B. Behavious has been
corrected.

Also handle the case when cycling between A & B but you focus C though
some other means. Focusing C will cause the window cache to be cleared
and the cycling to reset. Example: A B A C A ...