Making window cloning more of an extensions
Improved useability of expirimental window-cloning. In particular
improvements were made to how windows in different workspaces(with
different mapped states) are handled to avoid insible windows and
infinite swaping between parent and child

Removed the window-cloning related methods from the core code base. It
is now a complete add-on.
Also added the effectiveID field to WindowInfo such the clones can be
the same as the parent for the purpose of detecing state changes (ie for
retiling)

Title is now synced between clone and parent on update