Load window properties right before a window is first mapped
We load window properties when a window is first detected, but all the
window properties may not be set by this time. They will be set before
the window is mapped (according to the EWMH/ICCCM spec). We reloaded
properties whenever the client requested a mapping, but there could be
the case that we manually map the window on detection. For this case, we
manually re-load the properties.