Improved handleing of WM_SELECTION
It is intended for only one WM to be running at a time, but there was a
bug that caused a closing WM to clear the WM selection thus allowing
another WM to run without a problem. This has been fixed.

If another WM foribly steals our WM_SELECTION we die gracefully.

Added globals to forcfully steal WM_SELECTION and not to bother setting
it