Only look at map state when doing the initial scan
There was a bug where we would look at map state when a window was
detected instead of just at startup. Some windows, namely tray icons,
would have a postive map state which often implies they should be
mapped. However, the system tray causes them to not be mapped without
any notification so we are in an incorrect state. Now we only use map
state when initially starting to avoid this issue.