WM_TAKE_FOCUS atom has wrong value
WM_TAKE_FOCUS and the other manually created atoms have a value that is
different from other clients, which render the atom useless of
inter-client communication. THe problem was caused by specified a fixed
and incorrect length when creating the atom. This caused it to have
junk characters behind the '\0' which made them invisible.