Implemented a way to get the max number of master devices
createMaster device would fail if the device limit (40) was reached.
However this number is a hard constant and is not readily available to
the client. There is a merge request to XServer to post a property on
the root window, which we will now read. In the meantime we will return
the default 40.