Bug Fix: Detect state change with layout limit
There was a bad optimization in state.cpp that only looked compared
n windows (where n is the number of windows that will be tiled) to
determine if a retiled is needed. However this number depended on the
limit argument of the layout. So if a Workspace was already tiled, there
were more windows in it than the limit and a new tileable window was
added to the workspace, a retile would not be triggered. This caused the
new windows to have their default geometry until a retile was triggered
by other means (like another window changing state).