Removed profiling
The "at-scale" profiling is largly useless, because this product will
never be used "at-scale." It would be unexpected to have 100 windows, so
there is little point in testing at the hundreds of thousands. That said
there are probably low handing fruit in terms of performance issues, but
the real bottleneck is interacting with the XServer.

Removed link time optimization