Queue XEvents
We used to grab events from the XServer one at a time, which was slow
and led noticeable latency in the event of an event storm. Now we load
all possible events from the XServer and store them in a local fixed-size
buffer. This buffer will be used for the next events until it is empty.
At this point we query X more events and repeat.