Reduce delay for commitWindowMoveResize
When dragging windows, release the mouse button didn't immediatly stop
dragging. This was due to the possible storm of motion events that
prevented us from processing the ungrab quickly. This issue has been
largly mitigated when the preceeding commit, but could still be
possible. However, when dragging windows, the window is also updated by
the current mouse position instead of looking at montion event details.
So now the window position will be changed if we have received a new
event (of any type) since we last updated the position and that the
position has indeed changed;