Update state of visible workspaces first
The old method of arbitrarily updating workspaces in order was
problematic. Focus could be set to none if there was a point where no
windows were mapped. This also masked problems in window-clone which
checked for WindowMasks instead of Workspace visibility and led to
windows being swapped twice (effetely a no-op). Lastly, invisible
Workspaces can no longer be tiled. And there is now a special flag that
will to detect this case and mark the Workspace as dirty and it becomes
visible and tile it. This fixes a bug where if a window is move-backed
and forth between workspaces it will not be automatically tiled.

Improved debugging of state changes.