Prevent clients from spamming _NET_WM_USER_TIME
GTK clients would update _NET_WM_USER_TIME on a toplevel window every
time a key/button press/release was detected which was spam.
_NET_WM_USER_TIME_WINDOW was created to fix this problem by offloading
the updates to a dedicated window. However, GTK was using this window
because we didn't we only set _NET_SUPPORTING_WM_CHECK on the root
window instead of also setting it on the supporting window.