Changed how WindowMasks are synced with WM_STATE
The check is global and not on a per-window basis. As a result the SYNC
related WindowMasks hae been removed The global variables can determine
which masks will be synced and which type of application (normal app,
pager or other) can change them. Requests to change a non-synced state
will be ignored.
The README has also been updated with consequences of this decision and
potential workarounds if they are undesirable.

WM_ACTIONS better reflect actions clients can take that we won't always
ignore.

Don't apply global mask sync to non manageable windows