autoFocus will ignore windows in invisible workspace
autoFocus doesn't check if a window is MAPPED because it is intended to
be called from a XCB_MAP_NOTIFY. However, the window may have been
unmapped between when we received the event and now likly by being moved
to an invisible workspace. So this quick check was added before the
focus attempt.