state.cpp has returned to just being used for retiling
Other functionality is now implemented via rules and are enabled by
default. This allows for better control when switching workspaces with
STICK windows.

Focus may now be transferred from Desktop window to another when in the
Master stack when switching workspace.
Desktop rules are enabled by default.

When a window is unmapped focus will be transferred to the next available
window in the Master stack or root; The Workspace stack is not considered

TileWorkspace is applied after tiling