Basic gestures support
GestureEvents can now be generated from touch devices and function in a
similar way to regular X events. A list of GestureBindings can be used
to select interesting events and they will be triggered when a matching
event happens

The masks currently supported are GestureEnd, TouchEnd, TouchHold and
TouchMotion. Desired masks can be explicitly listened for

Store x coordinates with GestureEvent

An intended use case for GestureEvents is to trigger X11 events, so it
is helpful to to where on the screen (in pixels) an even occurred. The
last of a gesture is not stored in the event.