Improved threading
Switched from pthreads to std::thread (which may be backed by pthreads).
THe new version allows for more flexible methods and a better overall
interface.

Allow wakeup functions to be registered and called in
waitForAllThreadsToExit

Also introduced ThreadSignaler struct which provides a wrapper around
conditional wait and signaling for threads. This has replaced all the
busy sleeps with a blocking wait.