Redesign Options and fully test mpxmanager.cpp
Refactored Option class to use BoundFunction

mpxmanager has two classes of startup options. Those starting with "--"
and all others. In the former case, all options will be applied to the
current instance and only startup options and variables may be used. In
the latter case, options will be sent to a running instance and only one
may be used. The process will exit with the return status of the
command.

Removed send and set commands
Options have are lowercase and uses "-" instead of "_"
Remove arg tests