_NET_WM_MOVE_RESIZE now works when triggered by apps
Moving/resizing the window relies on grabbing the pointer to listen for
events, which is done automatically when in a chain, but not on client
messages. Now client messages will grab the pointer of the clients's
master and wait for the speciifed button to be relased before committing
the change.