Allow event to be run on chain end
In addition to having a function be triggered on chain start, one can be
triggered on chain end and abort.
Chains can end through normal termination or abortion, and seperate
functions can be run for each case. onEnd function will always be called
while the onAbort will be called depending on the new abort flag for
Chain::end

A new rule was also added to terminate chains when idle