Remove ALWAYS_ON_TOP/BOTTOM masks
The ALWAYS_ON_TOP and ALWAYS_ON_BOTTOM masks that listened
for configure events and raised/lower windows has been replaced. Now we
raise/lower non-transient windows relative to some override-redirect,
input-only windows whose sole purpose is to divide the windows into
layers. So now we wish to raise a normal window it will be raised to be just
below the boundary window for normal and ABOVE window