Container prototype
Containers where an attempt at a WM-level replacement for tabs. A window
was created that is treated as both WindowInfo and a Monitor (whose
geometry is the window's position). This allowed for nested workspaces
within a window. In comparison to application tabs, containers allowed
normal tiling operations to be applied to the window. A similar solution
could also be implemented with embedded windows (like tabbed), but now
there is no issue if the WM dies (along with the parent window) and the
same keybindings can be used to move windows to different contains as can
be used with workspaces.

There are still some open problems with containers. Mainly the
interaction between multiple workspaces occupying the same screen. The
stacking order is updated per workspace and windows are raised/lowered
globally (within their stacking layer). Ideally all the windows would
be adjacent to the container window always. This wasn't the biggest issue when
tiling since most layouts don't modify the stack, but it was very
prevalent when activating a window. Activating the container didn't
automatically cause its windows to be raised.