No longer read activeWorkspace property on start
ActiveWorkspace property never made since in the MPX mindset. Since this
property would have to change every time the active master changed. Its
not a problem to set it, but resuming from it is problematic. The
session extension provides a better, more comprehensive solution so this
responsibility has been delegated to it.