More natural behavior when cycling through findAndRaise
findAndRaise could be use to cycle through window matching an arbitrary
pattern but it behave un-intuitively once it cycled through windows. It
would then proceed to re-visit windows in the opposite order after it
started repeating. Now windows are cycled through in a loop so that if
there are N matching windows, the Nth switch will go back to the first.