[layouts] Don't tile windows to size -1
If a window had a size < the size of the border, when adjusting the size
for the border, the target size would go negative. However, the variable
was unsigned so it wrapped to max short value so the default min
width/height wasn't applied. Now values are casted to (signed) int when
checking if the min value of 1 should be applied or not.