Use the head of the Master window stack as the focused window
Master window stack was different than others in that the last window
represented the "top" of the stack. This felt more natural as this stack
was more of a "stack" (LIFO) than a general list like the Workspace
stack. However, this was extra complexity, so the master stack was made
to be in sync with the workspace stack. This shouldn't have an affect on
the end user experience.