Added extension to auto detect dock properties
Some suckless software (svkbd) supports docks somewhat. It will position
itself on the edge of the monitor and give itself the correct type, but
it won't set the EWMH dock properties. Now given the information we
have, we can derive the properties and treat it as a dock. That's what
this patch does.