Fixed bug that caused unmapped windows to be mapped on restart
Some clients private windows that they never map (and they aren't always
marked override redirect). We misread the ICCCM spec and assumed that if
the window had its wm hints set and the initial state property wasn't
withdrawn, then we should map the window on startup. We've correctness
the mistake and no ignore this bit as we don't support iconic state