Switched default keybindings from switchToWorkspace to activateWorkspace
Using switchToWorkspace had undesirable behavior when the master was on
a workspace with no active windows (focused on root) and switched to one
where it had active windows. There were attempts to update focus to the
head (or closed possible to head) of the master stack when a new window
was mapped, but this assumes the windows are mapped in a certain order
if done synchronously. Fixed this or doing in on (batch) map_notify
seemed like too much work, so activateWorkspace is being used instead
since its implementation circumvents this problem by explicitly mapping
the window inline.