Reduce window flashing when switching workspace
Previously we unmap windows arbitrarily which could lead to unmapping the
single visible window first which would cause the window right beneath
it to briefly flash on the screen before it too was deleted.
To address this, the VISIBILITY mask was re-added along with listening
for VISIBILITY events by default.