Load current geometry info on CLIENT_MAP_ALLOW
Before geometry was set as part of scan if the window was registered.
This meant that existing windows wouldn't have a valid geometry when
their CLIENT_MAP_ALLOW even triggered. Now geometry info is loaded as a
HIGHER_PRIORITY event so normal events can rely on the geometry always
being set by default.