Fix SIGBUS on arm when trying to set window position
setWindowPosistion accepts shorts and converts them to ints via the
copyTo helper. However it incorrectly passed in 1 to includeBuffer which
made the function attempt to copy an extra field. This caused an invalid
read/write that often resulted in a SIGBUS. The includeBuffer field has
just been removed since it was always supposed to be zero anyways.