Completely rework containers
Most of the complexity and inconvenience stemmed from the fact that the
underlying window as created by the WM. So it needed special handling
for when the WM restarted and one couldn't kill it normally without
killing the WM. Now one passes in a window and a monitor will be linked
to it (if it doesn't already exist). Creating the window is the job of
the client, but all they need to do is just open an X connection and
create a window and just wait for the window to be destroyed. Or just
use a normal window and make everything in that workspace fullscreen.