Switch to synchronous grabs
This is a fundemental change. Previously when issuing a grab, other
events were processed normally. Now the device will appear frozen until
we explicitly call unfreeze (which is now done after all events have
been processed). The driving force for this change, was to allow the
events to be replayed. Button1 was was already special cased, but now in
general, the client can abort processing of events and replay them as if
we never had the grab in the first place