Added cli to hide/unhide window
Windows with the HIDDEN_MASK can now be activated. To have them not be
activatable add the NO_ACTIVATE_MASK. The default bindings to hide
windows now also set this mask so their behavior is preserved.

This was done to differentiate the way one may hide windows. One way
(old) was to just banish a window that wasn't relevant. The new way is
to temporarily hide a window (like to change the effective set of
grabbed keys). In the latter case, one would want the normal
raiseOrRun/window activation code to find such a window. Now people can have both.
The new cli commands work in this new way.