Don't grab Button1 by default
Some websites (like slack) had a buggy UI when viewed via a recent
Firefox with default settings. When a mouse-hover window was opened, it
would immediately be closed. Presumably this was due to the temporary
focus change that happens when processing a grab. Passively listening
for the button press avoid this issue, so the defaults have been updated
to avoid this annoying behavior.