Tile workspace before activating a window
There has been a long standing bug where a newly activated window would
map itself before the workspace has been tiled. This meant that if the
window was originally at a different position it would false onscreen
before being tiled to the correct position. This was especially
noticeable with a multi-monitor config when a window briefly flashed on
the wrong monitor.